+++ packaging/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Now labels of dead players are hidden when switching biomes to not bloat up the UI.
 - Fixed player labels being kept after winning the game.
 - Fixed item ping labels moving to the map's root on pickup.
+- Fixed item ping's campfire icon being differently styled than the campfire indicator's icon.
 - Changed certain descriptions and enum values, to resolve some confusion about the placement of labels (off-screen / edge-of-screen).
 
 ## 1.1.2